Fix the sparsity similarity matrix computation
Previously the sparsity similarity matrix was built incorrectly, we were
using the `numpy.partition` function wrongly and in case of plateaux
(it's quite common to have the same value in the similarity matrix
repeated multiple times when using few features) the density of the
sparse matrix might not correctly represent what specified by the user.

Now we are relying on `numpy.argsort` that is slower than a regular sort
and is not multithreaded by default. In case we need a more efficient
implementation we can parallelize it with a cython extension.